roadmap generator server/requirements.txt
annotated-types
anyio
async-timeout
blinker
certifi
charset-normalizer
click
colorama
exceptiongroup
Flask
flask-cors
greenlet
h11
httpcore
httpx
idna
itsdangerous
Jinja2
jsonpatch
jsonpointer
langchain
langchain-core
langchain-ollama
langchain-text-splitters
langsmith
MarkupSafe
ollama
orjson
packaging
pip
pydantic
pydantic_core
python-dotenv
PyYAML
requests
requests-toolbelt
setuptools
sniffio
SQLAlchemy
tenacity
typing_extensions
typing-inspection
urllib3
Werkzeug
zstandard
